Split BIDS test data into separate module
